% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Annotate.R
\name{Annotate}
\alias{Annotate}
\title{Annotate}
\value{
An Annotate object
}
\description{
R6 class object used for generating and storing plots and other
    processed data. It is accessed internally to display data. Each top level
    list contains a different type of plot. Sublists are organised to store
    different plot versions.
    Functions for plotting (i.e. plot_actogram, plot_DAct) get data from
    fct_R6_Custom_tables object and generate a plot that is saved into the
    relative list.
    There should be a method to pass plots to an external container based on
    the handler.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{store}}{named list used as a keyed registry for all plot objects.
Use store_obj(key, object) and retrieve_obj(key) to read and write.
Keys used in the app: "ts_original", "ts_annotated",
"periodogram_total", "periodogram_faceted", "periodogram_averaged",
"period_barplot".}

\item{\code{actTable}}{Table containing data assembled using showData()}

\item{\code{metaTable}}{Table containing data assembled using showMeta()
showMeta}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Annotate-showMeta}{\code{Annotate$showMeta()}}
\item \href{#method-Annotate-showData}{\code{Annotate$showData()}}
\item \href{#method-Annotate-store_obj}{\code{Annotate$store_obj()}}
\item \href{#method-Annotate-retrieve_obj}{\code{Annotate$retrieve_obj()}}
\item \href{#method-Annotate-plot_Timeserie}{\code{Annotate$plot_Timeserie()}}
\item \href{#method-Annotate-plot_Timeserie_facet}{\code{Annotate$plot_Timeserie_facet()}}
\item \href{#method-Annotate-plot_periodogram}{\code{Annotate$plot_periodogram()}}
\item \href{#method-Annotate-plot_period_barplot}{\code{Annotate$plot_period_barplot()}}
\item \href{#method-Annotate-clone}{\code{Annotate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Annotate-showMeta"></a>}}
\if{latex}{\out{\hypertarget{method-Annotate-showMeta}{}}}
\subsection{Method \code{showMeta()}}{
function to create a table that arranges all the available
metadata in a table, and saves it in metaTable var
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Annotate$showMeta(env)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{env}}{App_settings environment to access myCleanMice object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
no return
showdata
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Annotate-showData"></a>}}
\if{latex}{\out{\hypertarget{method-Annotate-showData}{}}}
\subsection{Method \code{showData()}}{
function to create a table that arranges activity data in a
table, and saves it in actTable var
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Annotate$showData(env, id, miceList)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{env}}{App_settings environment to access myCleanMice object}

\item{\code{id}}{mouse Id value from the list of Ids or "All" to display all data}

\item{\code{miceList}}{list of all available mouse Ids}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
no return
store_obj
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Annotate-store_obj"></a>}}
\if{latex}{\out{\hypertarget{method-Annotate-store_obj}{}}}
\subsection{Method \code{store_obj()}}{
a function to store objects inside the Annotate R6 obj
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Annotate$store_obj(key, object)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{string to discriminate cases for storing file}

\item{\code{object}}{the object to be stored}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
no value returned
retrieve_obj
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Annotate-retrieve_obj"></a>}}
\if{latex}{\out{\hypertarget{method-Annotate-retrieve_obj}{}}}
\subsection{Method \code{retrieve_obj()}}{
retrieve a stored object by key
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Annotate$retrieve_obj(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{string key used when storing the object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the stored object, or NULL if the key does not exist
plot_Timeserie
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Annotate-plot_Timeserie"></a>}}
\if{latex}{\out{\hypertarget{method-Annotate-plot_Timeserie}{}}}
\subsection{Method \code{plot_Timeserie()}}{
a function to create Timeseries plots
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Annotate$plot_Timeserie(
  datatable,
  value,
  env,
  tmin = 0,
  tmax = 999.81,
  data1 = TRUE,
  data2 = FALSE,
  data3 = FALSE,
  timeformat = "Hours"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datatable}}{list contain all the data}

\item{\code{value}}{string to select data source: original, detrended or cleaned}

\item{\code{env}}{environment. App settings environment}

\item{\code{tmin}}{double. minimum time to be plotted}

\item{\code{tmax}}{double. maximum time to be plotted}

\item{\code{data1}}{Boolean. is dataset 1 present?}

\item{\code{data2}}{Boolean. is dataset 2 present?}

\item{\code{data3}}{Boolean. is dataset 3 present?}

\item{\code{timeformat}}{String. is the time in hours or posixt}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a ggplot object
plot_Timeserie_facet
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Annotate-plot_Timeserie_facet"></a>}}
\if{latex}{\out{\hypertarget{method-Annotate-plot_Timeserie_facet}{}}}
\subsection{Method \code{plot_Timeserie_facet()}}{
a function to create Timeseries plots stacked inside a grid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Annotate$plot_Timeserie_facet(
  datalist,
  value,
  env,
  tmin = 0,
  tmax = 999.81,
  data1 = TRUE,
  data2 = FALSE,
  data3 = FALSE,
  timeformat = "Hours"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datalist}}{list contain all the data}

\item{\code{value}}{string to select data source: original, detrended or cleaned}

\item{\code{env}}{environment. App settings environment}

\item{\code{tmin}}{double. minimum time to be plotted}

\item{\code{tmax}}{double. maximum time to be plotted}

\item{\code{data1}}{Boolean. is dataset 1 present?}

\item{\code{data2}}{Boolean. is dataset 2 present?}

\item{\code{data3}}{Boolean. is dataset 3 present?}

\item{\code{timeformat}}{String. is the time in hours or posixt}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a ggplot object
plot_periodogram
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Annotate-plot_periodogram"></a>}}
\if{latex}{\out{\hypertarget{method-Annotate-plot_periodogram}{}}}
\subsection{Method \code{plot_periodogram()}}{
Function to generate periodogram plots
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Annotate$plot_periodogram(
  method,
  plotType = c("Pertotal", "Perfaceted", "Peraveraged"),
  FunEnv
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{function to use from zeitgebr::periodogram function}

\item{\code{plotType}}{the different type of plots to generate. default to total and faceted}

\item{\code{FunEnv}}{AppSettings environment}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
function returns internally, no direct output

plot_period_barplot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Annotate-plot_period_barplot"></a>}}
\if{latex}{\out{\hypertarget{method-Annotate-plot_period_barplot}{}}}
\subsection{Method \code{plot_period_barplot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Annotate$plot_period_barplot(method, FunEnv)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{function used to run period analysis}

\item{\code{FunEnv}}{AppSettings environment}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
function returns internally
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Annotate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Annotate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Annotate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
