% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_loader.R
\name{load_data}
\alias{load_data}
\title{load_data}
\usage{
load_data(env, vessel = "Front-Left")
}
\arguments{
\item{env}{App_settings environment}

\item{vessel}{Vessel position when uploaded}
}
\description{
creates Clean_sample_data object and stores into myCleanSample list
}
