% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui_helpers.R
\name{overwrite_vessel}
\alias{overwrite_vessel}
\title{overwrite_vessel}
\usage{
overwrite_vessel(session, vessel = "Front-left")
}
\arguments{
\item{session}{session value}

\item{vessel}{name of vessel}
}
\value{
no return
}
\description{
overwrite_vessel
}
